+++ CHANGELOG.md
@@ -13,3 +13,8 @@
 
 ## 0.0.5
 - Documentation typos corrected.
+
+## 0.0.6
+- `App` has new static methods such as `goBack()`, `message(...)`, `alert(...)` etc. See the usages in the example.
+- New class: Field - represents a data entry field.
+- New class: `DataScreen` - for data entry. Contains helper methods for creating text field and date field. (More field types will be added in the future). See the usages in the example.